requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
SQLAlchemy==2.0.7
psycopg2-binary==2.9.5
python-dotenv==1.0.0
Werkzeug==2.2.3
pytest==7.3.1
